refractoring a bit more, swapped double for bigdecimal, added for sb building

+++ src/main/java/org/example/objects/RentAndInsurance.java
@@ -4,16 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter @Setter @AllArgsConstructor
 public class RentAndInsurance {
 
-    private double fullyChargedRent;
-    private double fullyChargedInsurance;
+    private BigDecimal fullyChargedRent;
+    private BigDecimal fullyChargedInsurance;
 
-    private double discountedRent;
-    private double discountedInsurance;
+    private BigDecimal discountedRent;
+    private BigDecimal discountedInsurance;
 
-    private double totalRentPaid;
-    private double totalInsurancePaid;
-    private double total;
+    private BigDecimal totalRentPaid;
+    private BigDecimal totalInsurancePaid;
+    private BigDecimal total;
 }
